Move AvatarFileInterceptor to shared
